feat(CUST-197)refactored consumer creation.consumer params separated by different structs.fix

+++ pkg/nats/js_consumer_workerpool_pull_type_chan.go
@@ -103,7 +103,7 @@ func NewJsPullTypeConsumerWorkersPool(logger *zap.Logger,
 		pullSubscriber: pullSubscriber,
 	}
 
-	for i := uint16(0); i < uint16(consumerCfg.GetWorkersCount()); i++ {
+	for i := uint32(0); i < consumerCfg.GetWorkersCount(); i++ {
 		ww := &jsConsumerWorkerWrapper{
 			msgChannel:       msgChannel,
 			stopWorkerChanel: make(chan bool),
